Move common into routes
- Moves `src/common` into `src/routes/common`
- Moves `src/common/middleware` to `src/middleware`
- `src/common` should only have route related components that should be shared between routes (such as common decorators and entities)
