+++ src/main.ts
@@ -2,7 +2,7 @@ import { VersioningType } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { DataSourceErrorFilter } from './common/filters/data-source-error.filter';
+import { DataSourceErrorFilter } from './routes/common/filters/data-source-error.filter';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
